feat: Add getFeaturedProduct route in Product module
The code changes in this commit add a new route, /featured, to the Product module in order to retrieve featured products. This functionality is implemented in the getFeaturedProduct function in the product.controller.ts file.

This commit follows the established convention of using a prefix to indicate the type of change, in this case, feat for a new feature.