feat: Update app.ts to allow CORS from client domain and create the Readme.Md
The code changes in this commit update the pp.ts file to allow CORS (Cross-Origin Resource Sharing) from the client domain. The cors middleware is modified to include the client domain https://flex-fit-hub-client.vercel.app in addition to http://localhost:5173.

This commit follows the established convention of using a prefix to indicate the type of change, in this case, feat for a new feature.